Attemt multiple tcp messages (does not work)

+++ fastmaprenderer/src/main/java/me/wcaleniewolny/ayaya/fastmaprenderer/fastmaprenderer/client/netty/MapNettyClient.java
@@ -10,6 +10,8 @@ import io.netty.channel.socket.nio.NioSocketChannel;
 import io.netty.handler.codec.LengthFieldBasedFrameDecoder;
 import io.netty.handler.codec.compression.JdkZlibDecoder;
 import io.netty.handler.codec.compression.ZlibWrapper;
+import io.netty.handler.logging.LogLevel;
+import io.netty.handler.logging.LoggingHandler;
 import me.wcaleniewolny.ayaya.fastmaprenderer.fastmaprenderer.client.RenderMetadata;
 
 public class MapNettyClient {
@@ -36,7 +38,7 @@ public class MapNettyClient {
                     @Override
                     protected void initChannel(Channel socketChannel) {
                         ChannelPipeline pipeline = socketChannel.pipeline();
-                        pipeline.addLast("framer", new LengthFieldBasedFrameDecoder(Short.MAX_VALUE, 0, 4, 0, 4));
+                        pipeline.addLast("framer", new LengthFieldBasedFrameDecoder(Integer.MAX_VALUE, 0, 4, 0, 4));
                         pipeline.addLast("decompression", new JdkZlibDecoder(ZlibWrapper.GZIP));
                         pipeline.addLast("handler", new NettyDataHandler());
                     }
